fix bugs

+++ bootstrap.sh
@@ -21,14 +21,15 @@ TOBE_LINKED=(
     ".curlrc"
     ".screenrc"
     ".wgetrc"
+    ".php_cs"
 )
-
+count=0
 while [ "x${TOBE_LINKED[count]}" != "x" ]
 do
    count=$(( $count + 1 ))
    link=${TOBE_LINKED[count]}
    # echo $link
-   if [[ -e $link ]]; then
+   if [[ -e $link && ! -L ~/$link ]]; then
        ln -sf ${__CURRDIR__}/$link ~/$link
    fi
 done
@@ -45,15 +46,13 @@ ZPREZTORCS=(
     "zlogin"
     "zlogout"
 )
+zcount=0
 while [[ "x${ZPREZTORCS[zcount]}" != "x" ]]; do
     zcount=$(( $zcount + 1 ))
     rcfile=${ZPREZTORCS[zcount]}
     rcpath=$__CURRDIR__/.zprezto/runcoms/${ZPREZTORCS[zcount]}
     # echo $rcpath
-    if [[ -f $rcpath ]]; then
+    if [[ -f $rcpath && ! -L ~/.${rcfile} ]]; then
         ln -sf $rcpath ~/.${rcfile}
     fi
 done
-
-
-
